requirements.txt
cmake
face-recognition
pandas
opencv-python
tensorflow
streamlit
keras
opencv-python-headless
streamlit-webrtc
pillow